lingvo/app/requirements.txt
psycopg2==2.8.6
django==3.1.7
djangorestframework==3.12.2
requests
djoser==2.0.3
djangorestframework-simplejwt==4.6.0
